requirements.txt
flask==2.0.1
werkzeug==2.0.3
mysql-connector-python==8.0.28
requests==2.28.1
numpy==1.23.5 
python-dotenv